Allow >256 vectors to be used
None of the defaults use this many vectors, but the real tools support at least up to 512 vectors for a custom logic ic and the theoretical limit for my T48 is 32-bits. I don't need 4GB of test vectors (nor do I see anyone needing it-- it'd tie up the programmer for days), but exhaustively testing 10 input bits (1024 vectors) would be nice for this PLD I'm trying to figure out.

This is an int instead of an explicitly sized type because it's an int everywhere it's used.