Moved source code to src/.
The number of minipro's source files has grown to the point they really
need to be in their own source directory.  This was initially suggested
by Ali Atashrooz a year ago and I wanted to wait on it until a jump back
to Github could be done.  That switch back is on indefinite hold.  My
apologies, Ali.